# backend/radar-device/src/main/java/ma/enset/radardevice/models/NewData.java
package ma.enset.radardevice.models;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data @AllArgsConstructor @NoArgsConstructor
public class NewData{
    private Long radarId;
    private String rn;
    private double vehicleSpeed;
}