# backend/radar-device/src/main/java/ma/enset/radardevice/models/Radar.java
package ma.enset.radardevice.models;

import lombok.Data;

@Data
public class Radar {
    private Long id;
    private String name;
    private int maxSpeed;
    private double longitude;
    private double latitude;
}